Split MyWorkWorkspace.tsx into component files
frontend/src/features/my-work/MyWorkWorkspace.tsx (1098 lines) shrinks to
~630 - ContinueWhereLeftOff, SavedSystemCard, PlanCard, EmptyPanel, and
TelemetrySection move to frontend/src/features/my-work/components/, one
file each, verbatim. LabelToggle moves with SavedSystemCard and
MetricCard/formatCompactEventCounts move with TelemetrySection (their
only consumers), staying module-private. No logic changes, no barrel
file, no changes outside frontend/src/features/my-work/ - App.tsx's
import of MyWorkWorkspace is unaffected. See
docs/superpowers/specs/2026-08-06-split-mywork-workspace-components-design.md.
